Warn when extracted/scriptlet metadata is ignored
