+++ tests/unit/pluginhandler/test_pluginhandler.py
@@ -15,13 +15,10 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from collections import OrderedDict
-import contextlib
 import copy
-import functools
 import os
 import shutil
 import stat
-import subprocess
 import tempfile
 from unittest.mock import (
     call,
@@ -30,9 +27,7 @@ from unittest.mock import (
     patch,
 )
 
-import testtools
 from testtools.matchers import Contains, Equals, FileExists, Not
-from testscenarios.scenarios import multiply_scenarios
 
 import snapcraft
 from . import mocks
@@ -827,100 +822,6 @@ class NextLastStepTestCase(unit.TestCase):
 
         self.assertThat(self.handler.last_step(), Equals('prime'))
         self.assertThat(self.handler.next_step(), Equals(None))
-
-
-class ScriptletSetVersionTestCase(unit.TestCase):
-    def test_set_version_in_pull(self):
-        handler = self.load_part('test_part', part_properties={
-            'override-pull': 'snapcraftctl set-version override-version'
-        })
-
-        handler.pull()
-        metadata = handler.get_pull_state().scriptlet_metadata
-        self.assertThat(metadata.get_version(), Equals('override-version'))
-
-    def test_set_version_in_build(self):
-        handler = self.load_part('test_part', part_properties={
-            'override-build': 'snapcraftctl set-version override-version'
-        })
-
-        handler.pull()
-        handler.build()
-        metadata = handler.get_build_state().scriptlet_metadata
-        self.assertThat(metadata.get_version(), Equals('override-version'))
-        self.assertFalse(handler.get_pull_state().scriptlet_metadata)
-
-    def test_set_version_in_stage(self):
-        handler = self.load_part('test_part', part_properties={
-            'override-stage': 'snapcraftctl set-version override-version'
-        })
-
-        handler.pull()
-        handler.build()
-        handler.stage()
-        metadata = handler.get_stage_state().scriptlet_metadata
-        self.assertThat(metadata.get_version(), Equals('override-version'))
-        self.assertFalse(handler.get_pull_state().scriptlet_metadata)
-        self.assertFalse(handler.get_build_state().scriptlet_metadata)
-
-    def test_set_version_in_prime(self):
-        handler = self.load_part('test_part', part_properties={
-            'override-prime': 'snapcraftctl set-version override-version'
-        })
-
-        handler.pull()
-        handler.build()
-        handler.stage()
-        handler.prime()
-        metadata = handler.get_prime_state().scriptlet_metadata
-        self.assertThat(metadata.get_version(), Equals('override-version'))
-        self.assertFalse(handler.get_pull_state().scriptlet_metadata)
-        self.assertFalse(handler.get_build_state().scriptlet_metadata)
-        self.assertFalse(handler.get_stage_state().scriptlet_metadata)
-
-
-class ScriptletSetVersionErrorTestCase(unit.TestCase):
-
-    scriptlet_scenarios = [
-        ('override-pull', {'override_pull': 'snapcraftctl set-version 1'}),
-        ('override-build', {'override_build': 'snapcraftctl set-version 2'}),
-        ('override-stage', {'override_stage': 'snapcraftctl set-version 3'}),
-        ('override-prime', {'override_prime': 'snapcraftctl set-version 4'}),
-    ]
-
-    scenarios = multiply_scenarios(scriptlet_scenarios, scriptlet_scenarios)
-
-    def test_set_version_multiple_times(self):
-        part_properties = {}
-        with contextlib.suppress(AttributeError):
-            part_properties['override-pull'] = self.override_pull
-        with contextlib.suppress(AttributeError):
-            part_properties['override-build'] = self.override_build
-        with contextlib.suppress(AttributeError):
-            part_properties['override-stage'] = self.override_stage
-        with contextlib.suppress(AttributeError):
-            part_properties['override-prime'] = self.override_prime
-
-        # A few of these test cases result in only one of these scriptlets
-        # being set. In that case, we actually want to double them up (i.e.
-        # call set-version twice in the same scriptlet), which should still be
-        # an error.
-        if len(part_properties) == 1:
-            for key, value in part_properties.items():
-                part_properties[key] += '\n{}'.format(value)
-
-        handler = self.load_part('test_part', part_properties=part_properties)
-
-        with testtools.ExpectedException(errors.ScriptletRunError):
-            silent_popen = functools.partial(
-                subprocess.Popen, stdout=subprocess.DEVNULL,
-                stderr=subprocess.DEVNULL)
-
-            with patch('subprocess.Popen', wraps=silent_popen):
-                handler.pull()
-                handler.build()
-                handler.stage()
-                handler.prime()
 
 
 class StateBaseTestCase(unit.TestCase):
@@ -2721,26 +2622,3 @@ class CleanBuildTestCase(unit.TestCase):
 
         # Make sure the install directory is gone
         self.assertFalse(os.path.exists(handler.plugin.installdir))
-
-
-class ScripletTestCase(unit.TestCase):
-
-    def test_run_prepare_scriptlet(self):
-        handler = self.load_part(
-            'test-part', part_properties={'prepare': 'touch prepare'})
-
-        handler.build()
-
-        before_build_file_path = os.path.join(handler.plugin.build_basedir,
-                                              'prepare')
-        self.assertThat(before_build_file_path, FileExists())
-
-    def test_run_install_scriptlet(self):
-        handler = self.load_part(
-            'test-part', part_properties={'install': 'touch install'})
-
-        handler.build()
-
-        after_build_file_path = os.path.join(handler.plugin.build_basedir,
-                                             'install')
-        self.assertThat(after_build_file_path, FileExists())
