+++ tests/unit/test_meta.py
@@ -521,6 +521,9 @@ class CreateMetadataFromSourceTestCase(CreateMetadataFromSourceBaseTestCase):
         self.assertThat(raised.part, Equals('wrong-part'))
 
     def test_metadata_doesnt_overwrite_specified(self):
+        fake_logger = fixtures.FakeLogger(level=logging.WARNING)
+        self.useFixture(fake_logger)
+
         def _fake_extractor(file_path):
             return extractors.ExtractedMetadata(
                 summary='extracted summary',
@@ -536,6 +539,13 @@ class CreateMetadataFromSourceTestCase(CreateMetadataFromSourceBaseTestCase):
         self.assertThat(y['summary'], Equals(self.config_data['summary']))
         self.assertThat(
             y['description'], Equals(self.config_data['description']))
+
+        # Verify that we warn that the YAML took precedence over the extracted
+        # metadata for summary and description
+        self.assertThat(fake_logger.output, Contains(
+            "The 'description' and 'summary' properties are specified in "
+            "adopted info as well as the YAML: taking the properties from the "
+            "YAML"))
 
     def test_metadata_with_unexisting_icon(self):
         def _fake_extractor(file_path):
@@ -659,6 +669,9 @@ class ScriptletsMetadataTestCase(CreateMetadataFromSourceBaseTestCase):
         self.assertThat(generated['version'], Equals('override-version'))
 
     def test_scriptlets_no_overwrite_existing_property(self):
+        fake_logger = fixtures.FakeLogger(level=logging.WARNING)
+        self.useFixture(fake_logger)
+
         del self.config_data['parts']['test-part']['parse-info']
         self.config_data['parts']['test-part']['override-prime'] = (
             'snapcraftctl set-version override-version')
@@ -666,6 +679,12 @@ class ScriptletsMetadataTestCase(CreateMetadataFromSourceBaseTestCase):
         generated = self.generate_meta_yaml(build=True)
 
         self.assertThat(generated['version'], Equals('test-version'))
+
+        # Since the specified version took precedence over the scriptlet-set
+        # version, verify that we warned
+        self.assertThat(fake_logger.output, Contains(
+            "The 'version' property is specified in adopted info as well as "
+            "the YAML: taking the property from the YAML"))
 
     def test_scriptlets_overwrite_extracted_metadata(self):
         del self.config_data['version']
